flatten-bom (#180)
A new mojo that helps you to maintain (and synchronize) with published (but bloated) BOMs.

```
mvn toolbox:flatten-bom -Dbom=g:a:v
```

And outputs flat BOM, that is in fact "effective BOM" with all conflicts (and recursion, most importantly) removed. The BOM being flattened must be resolvable (installed or deployed).

Optionally, you can specify `-Dgav=g:a:v` and output will be a BOM POM with provided coordinates.

Example invocation:
https://gist.github.com/cstamas/9fa43d43602e4fd9f28787e8f3cadd73