Bug: versions went for all effective deps
My pet peeve: your POM contains very few depMgt entries,
but one of them is import for Quarkus BOM. In this case
you end up with screens and screens of managed deps (and
those being checked), but in fact, you should not care
about them as all you care is the "stack" (BOM you import).

This fixes it.

Fixes #255