Mojo Concurrency Fix
The toolbox classes are a bit tricky: they are Maven Mojos, but also
also CLI/Picocli commands as well. By default, they are "one-shot"
and should run as such.

But, in some special cases (REPL, or mvnsh), their context should
be carried on, but not always.

The CONTEXT was `static` to carry on contextual information, but
this was wrong, as it is needed only in afore mentioned cases,
but not when running as Mojo (where it in fact introduces concurrency
issues).

Fix is that two places (REPL and mvnsh) explicitly carries over
context, while in case of Mojo they are using own context.