Contextual filtering (#309)
All the version selector/matcher combo were wrong (almost, some were ok).
The problem: matcher was `noSnapshotsAndPreviews()` and selector was `last()`.

New setup: matcher is `any()` and selector is `contextualSnapshotsAndPreviews()`

The new selector is now selecting based on initial state:
* IF initial state is snapshot version, no filtering applied
* IF initial state is preview version, filters out snaphots
* IF initial state is not preview nor snapshot, noPreviewAndSnapshot() is applied

The new selector accepts 0..1 arguments, with 0 arg defaults to `major()` selector,
but with 1 argument it may be further customized.